+++ src/ergo/network/ergoNetwork.ts
@@ -10,15 +10,15 @@ import {
   TxInput,
   TxOutput,
 } from './types';
-import { JsonBI } from './parser';
+import JsonBigInt from '@rosen-bridge/json-bigint';
 import { ergoTreeToBase58Address } from '../../utils/utils';
 import { ConnectionError } from '../../errors/errors';
 import { getConfig } from '../../config/config';
 import { ExplorerBox, ErgoAssetInfo } from '../network/types';
 import { MAX_API_LIMIT } from '../../config/constants';
-import { CallbackLoggerFactory } from '@rosen-bridge/callback-logger';
+import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 
-const logger = CallbackLoggerFactory.getInstance().getLogger(import.meta.url);
+const logger = DefaultLogger.getInstance().child(import.meta.url);
 
 export const explorerApi = axios.create({
   baseURL: getConfig().general.explorerUrl,
@@ -55,7 +55,7 @@ export class ErgoNetwork {
     return explorerApi
       .get<ExplorerBoxes>(`/api/v1/boxes/unspent/byErgoTree/${tree}`, {
         params: { offset: offset, limit: limit },
-        transformResponse: (data) => JsonBI.parse(data),
+        transformResponse: (data) => JsonBigInt.parse(data),
       })
       .then((res) => res.data);
   };
@@ -128,7 +128,7 @@ export class ErgoNetwork {
     while (offset < total && remaining()) {
       const boxes = await this.getBoxesForAddress(tree, offset, MAX_API_LIMIT);
       const ergoBoxes = wasm.ErgoBoxes.from_boxes_json(
-        boxes.items.map((box) => JsonBI.stringify(box).toString())
+        boxes.items.map((box) => JsonBigInt.stringify(box).toString())
       );
       logger.debug(
         `total boxes: ${total}, offset: ${offset}, number of current boxes: ${boxes.items.length}`
@@ -248,7 +248,7 @@ export class ErgoNetwork {
           inputs.forEach((input) => {
             const box =
               outputs.length > 0
-                ? wasm.ErgoBox.from_json(JsonBI.stringify(outputs[0]))
+                ? wasm.ErgoBox.from_json(JsonBigInt.stringify(outputs[0]))
                 : undefined;
             memPoolBoxesMap.set(input.boxId, box);
           });