+++ src/ergo/transaction/queue.ts
@@ -6,10 +6,9 @@ import { base64ToArrayBuffer } from '../../utils/utils';
 import { WatcherUtils } from '../../utils/watcherUtils';
 import { getConfig } from '../../config/config';
 import { Transaction } from '../../api/Transaction';
-import { CallbackLoggerFactory } from '@rosen-bridge/callback-logger';
-import { TxStatus } from '../../database/entities/observationStatusEntity';
+import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 
-const logger = CallbackLoggerFactory.getInstance().getLogger(import.meta.url);
+const logger = DefaultLogger.getInstance().child(import.meta.url);
 
 export class Queue {
   database: WatcherDataBase;