+++ src/jobs/minimumFee.ts
@@ -1,8 +1,8 @@
-import { CallbackLoggerFactory } from '@rosen-bridge/callback-logger';
+import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 import MinimumFeeHandler from '../utils/MinimumFeeHandler';
 import { getConfig } from '../config/config';
 
-const logger = CallbackLoggerFactory.getInstance().getLogger(import.meta.url);
+const logger = DefaultLogger.getInstance().child(import.meta.url);
 
 /**
  * runs MinimumFee update job