+++ src/jobs/transactionQueue.ts
@@ -2,9 +2,9 @@ import { Queue } from '../ergo/transaction/queue';
 import { WatcherDataBase } from '../database/models/watcherModel';
 import { WatcherUtils } from '../utils/watcherUtils';
 import { getConfig } from '../config/config';
-import { CallbackLoggerFactory } from '@rosen-bridge/callback-logger';
+import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 
-const logger = CallbackLoggerFactory.getInstance().getLogger(import.meta.url);
+const logger = DefaultLogger.getInstance().child(import.meta.url);
 
 let transactionQueue: Queue;