+++ src/transactions/detachWID.ts
@@ -6,9 +6,9 @@ import { ChangeBoxCreationError, NotEnoughFund } from '../errors/errors';
 import { TxType } from '../database/entities/txEntity';
 import { TransactionUtils } from '../utils/watcherUtils';
 import { getConfig } from '../config/config';
-import { CallbackLoggerFactory } from '@rosen-bridge/callback-logger';
+import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 
-const logger = CallbackLoggerFactory.getInstance().getLogger(import.meta.url);
+const logger = DefaultLogger.getInstance().child(import.meta.url);
 
 class DetachWID {
   /**