+++ src/utils/scanner.ts
@@ -11,7 +11,7 @@ import {
   DogeEsploraObservationExtractor,
   DogeRpcObservationExtractor,
 } from '@rosen-bridge/bitcoin-observation-extractor';
-import { CallbackLoggerFactory } from '@rosen-bridge/callback-logger';
+import { DefaultLogger } from '@rosen-bridge/abstract-logger';
 import { ErgoObservationExtractor } from '@rosen-bridge/ergo-observation-extractor';
 import {
   CardanoBlockFrostObservationExtractor,
@@ -70,31 +70,21 @@ import {
   createErgoExplorerNetworkConnectorManager,
 } from './networkConnectorManagers';
 
+const logger = DefaultLogger.getInstance().child(import.meta.url);
+
 /**
  * Creates loggers for scanners and extractors
  * @returns loggers object
  */
 const createLoggers = () => ({
-  commitmentExtractorLogger: CallbackLoggerFactory.getInstance().getLogger(
-    'commitment-extractor'
-  ),
-  eventTriggerExtractorLogger: CallbackLoggerFactory.getInstance().getLogger(
-    'event-trigger-extractor'
-  ),
-  observationExtractorLogger: CallbackLoggerFactory.getInstance().getLogger(
-    'observation-extractor'
-  ),
-  permitExtractorLogger:
-    CallbackLoggerFactory.getInstance().getLogger('permit-extractor'),
-  plainExtractorLogger:
-    CallbackLoggerFactory.getInstance().getLogger('plain-extractor'),
-  scannerLogger: CallbackLoggerFactory.getInstance().getLogger('scanner'),
-  observationScannerLogger: CallbackLoggerFactory.getInstance().getLogger(
-    'observation-scanner'
-  ),
-  collateralExtractorLogger: CallbackLoggerFactory.getInstance().getLogger(
-    'collateral-extractor'
-  ),
+  commitmentExtractorLogger: logger.child('commitmentExtractor'),
+  eventTriggerExtractorLogger: logger.child('eventTriggerExtractor'),
+  observationExtractorLogger: logger.child('observationExtractor'),
+  permitExtractorLogger: logger.child('permitExtractor'),
+  plainExtractorLogger: logger.child('plainExtractor'),
+  scannerLogger: logger.child('scanner'),
+  observationScannerLogger: logger.child('observationScanner'),
+  collateralExtractorLogger: logger.child('collateralExtractor'),
 });
 
 const loggers = createLoggers();
@@ -298,14 +288,20 @@ class CreateScanner {
       config.scannerType,
       config.address,
       undefined, // Token constraint not needed
-      loggers.plainExtractorLogger
+      loggers.plainExtractorLogger,
+      false
     );
+
     const collateralExtractor = new CollateralExtractor(
+      dataSource,
       Constants.COLLATERAL_EXTRACTOR_NAME,
       rosenConfig.AWC,
-      rosenConfig.watcherCollateralAddress,
-      dataSource,
-      config.explorerUrl,
+      {
+        active: true,
+        type: config.scannerType,
+        url: networkUrl,
+        address: rosenConfig.watcherCollateralAddress,
+      },
       loggers.collateralExtractorLogger
     );
     this.ergoScanner.registerExtractor(commitmentExtractor);